Station and build errors fixes from maritin traffic branch

+++ backend/database/schema.sql
@@ -282,7 +282,7 @@ CREATE TABLE IF NOT EXISTS stations (
     last_seen TIMESTAMPTZ,
     merged_into_id INTEGER REFERENCES stations(id) ON DELETE SET NULL,
     reach_coverage DOUBLE PRECISION,
-    building_count DOUBLE PRECISION,             -- absolute count of bike_path_buildings within 150m (computed at ingestion)
+    building_count INTEGER,                      -- absolute count of bike_path_buildings within 150m (computed at ingestion)
     UNIQUE (citybikes_network_id, station_id)
 );
 CREATE INDEX IF NOT EXISTS idx_stations_merged_into ON stations(merged_into_id);
